Waterfall: accurate timespan
Draw same line repeatedly if needed for short waterfall timespans or
slow FFT update rates.

+++ Waterfall.h
@@ -278,7 +278,7 @@ public:
 
     int     getNearestPeak(QPoint pt);
     void    setWaterfallSpan(quint64 span_ms);
-    quint64 getWfTimeRes(void);
+    double  getWfTimeRes(void);
     void    setFftRate(int rate_hz);
     void    clearWaterfall(void);
     bool    saveWaterfall(const QString & filename) const;
@@ -300,7 +300,7 @@ public:
 
     NamedChannelSetIterator channelCBegin() const;
     NamedChannelSetIterator channelCEnd() const;
-    
+
 signals:
     void newCenterFreq(qint64 f);
     void newDemodFreq(qint64 freq, qint64 delta); /* delta is the offset from the center */
@@ -501,9 +501,9 @@ private:
     qint64      m_upperFreqLimit    = 300000000;
 
     // Waterfall averaging
-    quint64     tlast_wf_ms;        // last time waterfall has been updated
-    quint64     msec_per_wfline;    // milliseconds between waterfall updates
-    quint64     wf_span;            // waterfall span in milliseconds (0 = auto)
+    double      tlast_wf_ms;        // last time waterfall has been updated
+    double      msec_per_wfline;    // milliseconds between waterfall updates
+    double      wf_span;            // waterfall span in milliseconds (0 = auto)
     int         fft_rate;           // expected FFT rate (needed when WF span is auto)
     int         m_expectedRate;
